+++ docs/netbox-extensions/changes/changelog.md
@@ -1,4 +1,9 @@
-# Change Log
+---
+title: Change Log
+tags:
+  - netbox-community
+  - netbox-enterprise
+---# Change Log
 
 ## v0.2.0